Refactor profile dashboard
