+++ cicero-dashboard/components/profile/types.ts
@@ -1,0 +1,18 @@
+export interface Stat {
+  label: string;
+  value: number | string;
+}
+
+export interface Activity {
+  id: string;
+  text: string;
+  date: string;
+}
+
+export interface ProfileData {
+  avatar: string;
+  name: string;
+  email: string;
+  stats: Stat[];
+  activity: Activity[];
+}
